[refactor] Add RedisMemo::Util
Pull Request Branch: donaldong/refactor_add_redismemo_util
Pull Request Link: https://github.com/chanzuckerberg/redis-memo/pull/71
